OvmfPkg: refine Platform GOP Policy
GOP driver has dependency on PlatformGopPolicy protocol, especially
the GetVbtData() interface, which we should return the corresponding VBT.

The GOP driver for IGD pass-thru needs the VBT data
in course of GPU display information initialization.
Currently it uses the VBT data provided by the board manufacturer.

But it is not convenient to build the VBT into OVMF.
At the same time, it requires customer
to fetch the VBT from the board manufacturer.

The VBT from host OpRegion is one choice.
So the GetVbtData() will try to parse the VBT from the passed Opregion.

The VBT data dumped from OpRegion is different compared with the native one,
as part of data has been stripped, but the checksum hasn't been updated.
So we need to re-calculate the checksum of VBT to make GOP driver happy.

Tracked-On: projectacrn/acrn-hypervisor#6603

Signed-off-by: Liu,Junming <junming.liu@intel.com>
Signed-off-by: Xiong Zhang <xiong.y.zhang@intel.com>
Signed-off-by: Colin Xu <colin.xu@intel.com>
Reviewed-by: Zhao Yakui <yakui.zhao@intel.com>
Reviewed-by: Yu Wang <yu1.wang@intel.com>