OVMF:ACRN:PCI: Try to load ROM image for the PCI device with PCI_ROM
The OVMF provides one mechanism to load the specific EFI image from PCI Rom
Bar. This can help ACRN to load the specific EFI image for the PCI device
instead of integrating it into the OVMF.fd.
PCI scan uses the light mechanism to enumerate the PCI devices. And it will
skip the parsing of PCI rom.

This will try to load the PCI rom in course of light PCI_scan enumeration.

Tracked-On: #8175
Signed-off-by: Zhao Yakui <yakui.zhao@intel.com>